Rewrote drawing functions to use new memory management system
Switched to doing memory management myself instead of using malloc() last commit, this is the actual implementation of that. Rewrote RenderQueueItems to be better for 3D and more flexible for 2D, at the cost of a bit of memory space. Also removed old code for the old render queue system.