PWM sync signals update, "merging" pwm-sync-x branches
After extensive testing and validation, the pwm-based sync stuff is finally ready. I "merged" it into the main code (the testing branches are based on such an old version of my code that I couldn't merge via pull request), and it "should" all work. It's the same parameters copied over.