SDK-ification complete
I decided to remove all direct register writes and replace them with calls to the pico sdk. It makes the code a bit more readable for other people using the pico who may not be used to direct writes. It's probably a bit slower, but since the calls happen in places that don't really matter I'm fine with it.