PCB v2 done
This update removed most of the stuff on the board -- the massive controller multiplexer is now replaced with an ATTiny24 on the controller side and an I2C bus. The buffer for the audio circuit is also gone. The board now supports surface mounting the pico, and is significantly smaller.